requirements.txt
selenium==4.18.1
beautifulsoup4==4.12.3
webdriver-manager==4.0.1
python-dotenv==1.0.1
